# fitnessWebApp-UI/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Andrea Barutta",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Andrea Barutta",
      "url": "https://barutta02.github.io/Barutta_portfolio/",
      "email": "andreabarutta123@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://some-url.com/api/v1",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/trainings": {
      "post": {
        "tags": [
          "Train"
        ],
        "operationId": "saveAllenamento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllenamentoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllenamentoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/trainingexercise": {
      "post": {
        "tags": [
          "TrainExercise"
        ],
        "operationId": "saveAllenamentoEsercizio",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllenamentoEsercizioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllenamentoEsercizioResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/trainings/{allenamento-id}": {
      "get": {
        "tags": [
          "Train"
        ],
        "operationId": "findAllenamentoById",
        "parameters": [
          {
            "name": "allenamento-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllenamentoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/trainings/creator": {
      "get": {
        "tags": [
          "Train"
        ],
        "operationId": "findAllAllenamentoByCreator",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAllenamentoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/trainingexercise/{allenamento-nome}": {
      "get": {
        "tags": [
          "TrainExercise"
        ],
        "operationId": "findByAllenamentoNome",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "allenamento-nome",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAllenamentoEsercizioResponse"
                }
              }
            }
          }
        }
      }
    },
    "/trainingexercise/no_pagination/{allenamento-nome}": {
      "get": {
        "tags": [
          "TrainExercise"
        ],
        "operationId": "findByAllenamentoNomeNoPagination",
        "parameters": [
          {
            "name": "allenamento-nome",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AllenamentoEsercizioResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/trainings/{allenamento-nome}": {
      "delete": {
        "tags": [
          "Train"
        ],
        "operationId": "deleteAllenamento",
        "parameters": [
          {
            "name": "allenamento-nome",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/trainingexercise/{allenamentoEsercizio-id}": {
      "delete": {
        "tags": [
          "TrainExercise"
        ],
        "operationId": "deleteAllenamentoEsercizio",
        "parameters": [
          {
            "name": "allenamentoEsercizio-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AllenamentoRequest": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "durata_in_ore": {
            "minimum": 0,
            "type": "number",
            "format": "float"
          }
        }
      },
      "AllenamentoResponse": {
        "required": [
          "creator_username",
          "description",
          "durata_in_ore",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "durata_in_ore": {
            "type": "number",
            "format": "float"
          },
          "creator_username": {
            "type": "string"
          }
        }
      },
      "AllenamentoEsercizioRequest": {
        "required": [
          "id_allenamento",
          "id_esercizio",
          "index",
          "recupero",
          "ripetizioni",
          "serie"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "id_allenamento": {
            "type": "integer",
            "format": "int64"
          },
          "id_esercizio": {
            "type": "integer",
            "format": "int64"
          },
          "index": {
            "type": "integer",
            "format": "int32"
          },
          "ripetizioni": {
            "type": "integer",
            "format": "int32"
          },
          "serie": {
            "type": "integer",
            "format": "int32"
          },
          "recupero": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AllenamentoEsercizioResponse": {
        "required": [
          "allenamento_id",
          "esercizio_id",
          "id",
          "index",
          "recupero",
          "ripetizioni",
          "serie"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "allenamento_id": {
            "type": "integer",
            "format": "int64"
          },
          "esercizio_id": {
            "type": "integer",
            "format": "int64"
          },
          "creator": {
            "type": "string"
          },
          "index": {
            "type": "integer",
            "format": "int32"
          },
          "ripetizioni": {
            "type": "integer",
            "format": "int32"
          },
          "serie": {
            "type": "integer",
            "format": "int32"
          },
          "recupero": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PageResponseAllenamentoResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllenamentoResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PageResponseAllenamentoEsercizioResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllenamentoEsercizioResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}